@@ -4,112 +4,110 @@ namespace ExeGeseIT\DoctrineQuerySearchHelper;
 
 use Doctrine\DBAL\Query\QueryBuilder as QueryBuilderDBAL;
 use Doctrine\ORM\QueryBuilder;
-use InvalidArgumentException;
 
 /**
  * Description of QuerySearchFilterFactory
  *
  * @author Jean-Claude GLOMBARD <jc.glombard@gmail.com>
+ *
+ * @phpstan-import-type TSearchvalue from SearchHelper
  */
 class QueryClauseBuilder
 {
+    private QueryBuilder|QueryBuilderDBAL $qb;
+
     /**
-     * @var \Doctrine\ORM\QueryBuilder | \Doctrine\DBAL\Query\QueryBuilder
+     * @var array<string, string>
      */
-    private $qb;
-    private $searchFields = [];
-    private $defaultLike = [];
-    
+    private array $searchFields = [];
+
+    /**
+     * @var array<string>
+     */
+    private array $defaultLike = [];
+
     private function __construct()
-    {}
-    
-    /**
-     * 
-     * @param \Doctrine\ORM\QueryBuilder | \Doctrine\DBAL\Query\QueryBuilder $qb
-     * @return self
-     * @throws InvalidArgumentException
-     */
-    public static function getInstance($qb): self
     {
-        if ( !($qb instanceof QueryBuilder || $qb instanceof QueryBuilderDBAL) ) {
-            throw new InvalidArgumentException( sprintf('$qb should be instance of %s or %s class', QueryBuilder::class, QueryBuilderDBAL::class) );
-        }
-        
+    }
+
+    public static function getInstance(QueryBuilder|QueryBuilderDBAL $qb): self
+    {
         $instance = new self();
         $instance->qb = $qb;
+
         return $instance;
     }
-    
-    
+
     /**
-     * 
-     * @param iterable $searchFields [searchKey => field]
-     * @return self
+     * @param array<string, string> $searchFields [searchKey => field]
      */
     public function setSearchFields(iterable $searchFields): self
     {
         foreach ($searchFields as $searchKey => $field) {
-            $this->searchFields[ $searchKey ] = $field;
+            $this->searchFields[$searchKey] = $field;
         }
+
         return $this;
     }
-    
+
     /**
      * If these searchKey appear in the $search array without any filter a LIKE filter is implicitly applied.
      * In other words, for such a searchKey, these two definitions are equivalent:
      *    SearchFilter::filter('default_like_searchkey') => 'foo',
      *    SearchFilter::like('default_like_searchkey') => 'foo',
-     * 
-     * @param iterable $likeFields [searchKey => field]
-     * @return self
+     *
+     * @param array<string, string> $likeFields [searchKey => field]
      */
     public function setDefaultLikeFields(iterable $likeFields): self
     {
         foreach ($likeFields as $searchKey => $field) {
-            $this->searchFields[ $searchKey ] = $field;
+            $this->searchFields[$searchKey] = $field;
             $this->defaultLike[] = $searchKey;
         }
+
         return $this;
     }
-    
-    public function getQueryBuilder(?iterable $search, ?string $paginatorSort): QueryBuilder|QueryBuilderDBAL
+
+    /**
+     * @ template TSearchvalue of array<int|string>|bool|int|string
+     * @param array<string, TSearchvalue|array<string, TSearchvalue>>|null $search
+     */
+    public function getQueryBuilder(?array $search, ?string $paginatorSort): QueryBuilder|QueryBuilderDBAL
     {
         $searchFilters = $this->getSearchFilters($search);
-        
+
         SearchHelper::initializeQbPaginatorOrderby($this->qb, $paginatorSort);
         SearchHelper::setQbSearchClause($this->qb, $searchFilters, $this->searchFields);
-        
+
         return $this->qb;
     }
-    
-    
+
     /**
-     * 
-     * @param iterable|null $search
-     * @return iterable
+     * @ template TSearchvalue of array<int|string>|bool|int|string
+     * @param array<string, TSearchvalue|array<string, TSearchvalue>>|null $search
+     * @return array<string, TSearchvalue|array<string, TSearchvalue>>
      */
-    private function getSearchFilters(?iterable $search): iterable
+    private function getSearchFilters(?array $search): array
     {
-        if ( empty($search) ) {
+        if (empty($search)) {
             return [];
         }
-        
-        if ( empty($this->defaultLike) ) {
+
+        if (empty($this->defaultLike)) {
             return $search;
         }
-        
+
         $_search = [];
         foreach ($search as $searchfilter => $value) {
-            
             $m = SearchFilter::decodeSearchfilter($searchfilter);
-            if ( empty($m['filter']) && in_array($m['key'], $this->defaultLike) ) {
-                $_search[ SearchFilter::like($m['key']) ] = ctype_print($value) ? trim($value) : $value;
+
+            if (empty($m['filter']) && in_array($m['key'], $this->defaultLike)) {
+                $_search[SearchFilter::like($m['key'])] = $value;
             } else {
-                $_search[ $searchfilter ] = $value;
+                $_search[$searchfilter] = $value;
             }
         }
-        
+
         return $_search;
     }
-            
 }
