@@ -9,37 +9,19 @@ namespace ExeGeseIT\DoctrineQuerySearchHelper;
  */
 class SearchFilter
 {
-    const EQUAL = '=';
-    const NOT_EQUAL = '!';
-    const LIKE = '%';
-    const NOT_LIKE = '!%';
-    const NULL = '_';
-    const NOT_NULL = '!_';
-    const GREATER = '>';
-    const GREATER_OR_EQUAL = '>=';
-    const LOWER = '<';
-    const LOWER_OR_EQUAL = '<=';
-    
-    const OR = '|';
-    const AND_OR = '&|';
-    
-    private static function getAlias(): iterable
-    {
-        yield self::EQUAL => [ '==' ];
-        yield self::NOT_EQUAL => [ '!=' ];
-        yield self::OR => [ '||' ];
-    }
-    
-    private static function getToken(): string
-    {
-        return bin2hex(random_bytes(15));
-    }
-    
-    private static function tokenize(bool $truly): string
-    {
-        return $truly ? '~' . self::getToken() : '';
-    }
-    
+    public const EQUAL = '=';
+    public const NOT_EQUAL = '!';
+    public const LIKE = '%';
+    public const NOT_LIKE = '!%';
+    public const NULL = '_';
+    public const NOT_NULL = '!_';
+    public const GREATER = '>';
+    public const GREATER_OR_EQUAL = '>=';
+    public const LOWER = '<';
+    public const LOWER_OR_EQUAL = '<=';
+    public const OR = '|';
+    public const AND_OR = '&|';
+
     public static function normalize(string $filter): string
     {
         $_filter = $filter;
@@ -49,218 +31,215 @@ class SearchFilter
                 break;
             }
         }
+
         return $_filter;
     }
-    
+
     /**
-     * 
-     * @param string $searchfilter
-     * @return array ['key' =>, 'filter =>]
+     * @return array{key: string, filter: string}
      */
     public static function decodeSearchfilter(string $searchfilter): array
     {
         $matches = [];
         preg_match('/(?P<filter>[^[:alnum:]]+)?(?P<key>[[:alnum:]][^~]*)/i', $searchfilter, $matches);
+
         return [
             'key' => $matches['key'] ?? '',
             'filter' => $matches['filter'] ?? '',
         ];
     }
-    
+
     /**
-     * 
-     * @param string $searchKey
-     * @param array $searchfilters
-     * @return bool
+     * @param array<string, string> $searchParameters
      */
     public static function hasFilteredKey(string $searchKey, array $searchParameters): bool
     {
         return null !== self::getFilteredKey($searchKey, $searchParameters);
     }
-    
+
     /**
      * Return <filteredkey> searchKey or null
-     * 
-     * @param string $searchKey
-     * @param array $searchParameters
-     * @return string|null
+     *
+     * @param array<string, array<int|string>|bool|int|string> $searchParameters
      */
     public static function getFilteredKey(string $searchKey, array $searchParameters): ?string
     {
         $hash = sprintf(':%s', implode(':', array_keys($searchParameters)));
-        $pattern = '/:(?P<filteredkey>(?:[^[:alnum:]])?' . $searchKey .'~[[:alnum:]]+)/i';
+        $pattern = '/:(?P<filteredkey>(?:[^[:alnum:]])?' . $searchKey . '~[[:alnum:]]+)/i';
         preg_match($pattern, $hash, $matches);
+
         return $matches['filteredkey'] ?? null;
     }
-    
+
     /**
      * Return <searchKey> $searchParameters value or null
-     * 
-     * @param string $searchKey
-     * @param array $searchParameters
-     * @return mixed
+     *
+     * @param array<string, array<int|string>|bool|int|string> $searchParameters
+     * @return array<int|string>|bool|int|string|null
      */
-     public static function getFilteredKeyValue(string $searchKey, array $searchParameters)
+    public static function getFilteredKeyValue(string $searchKey, array $searchParameters): string|int|array|bool|null
     {
         $filteredKey = self::getFilteredKey($searchKey, $searchParameters);
-        return empty($filteredKey) ? null : $searchParameters[ $filteredKey ];
+
+        return empty($filteredKey) ? null : $searchParameters[$filteredKey];
     }
-    
-    
+
     /**
      * ...WHERE 1
      *    {{ AND ( .. OR .. OR ..) }}
-     * 
-     * @return string
      */
     public static function andOr(): string
     {
         return self::AND_OR . self::getToken();
     }
+
     /**
      * ...WHERE 1
      *    {{ OR ( .. AND .. AND ..) }}
-     * 
-     * @return string
      */
     public static function or(): string
     {
         return self::OR . self::getToken();
     }
-    
-    
-    
+
     /**
      * isset($value)  ? => ...WHERE {{ searchKey = $value }}
      * !isset($value) ? => ...WHERE {{ 1 }}
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function filter(string $searchKey, bool $tokenize = true): string
     {
         return trim($searchKey) . self::tokenize($tokenize);
     }
+
     /**
      * ...WHERE 1
      *    {{ AND searchKey = $value }}
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function equal(string $searchKey, bool $tokenize = true): string
     {
         return self::EQUAL . trim($searchKey) . self::tokenize($tokenize);
     }
+
     /**
      * ...WHERE 1
      *    {{ AND searchKey <> $value }}
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function notEqual(string $searchKey, bool $tokenize = true): string
     {
         return self::NOT_EQUAL . trim($searchKey) . self::tokenize($tokenize);
     }
+
     /**
      * ...WHERE 1
      *    {{ AND searchKey LIKE $value }}
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function like(string $searchKey, bool $tokenize = true): string
     {
         return self::LIKE . trim($searchKey) . self::tokenize($tokenize);
     }
+
     /**
      * ...WHERE 1
      *    {{ AND searchKey NOT LIKE $value
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function notLike(string $searchKey, bool $tokenize = true): string
     {
         return self::NOT_LIKE . trim($searchKey) . self::tokenize($tokenize);
     }
+
     /**
      * ...WHERE 1
      *    {{ AND searchKey IS NULL
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function null(string $searchKey, bool $tokenize = true): string
     {
         return self::NULL . trim($searchKey) . self::tokenize($tokenize);
     }
+
     /**
      * ...WHERE 1
      *    {{ AND searchKey IS NOT NULL
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function notNull(string $searchKey, bool $tokenize = true): string
     {
         return self::NOT_NULL . trim($searchKey) . self::tokenize($tokenize);
     }
+
     /**
      * ...WHERE 1
      *    {{ AND searchKey > $value
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function greater(string $searchKey, bool $tokenize = true): string
     {
         return self::GREATER . trim($searchKey) . self::tokenize($tokenize);
     }
+
     /**
      * ...WHERE 1
      *    {{ AND searchKey >= $value
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function greaterOrEqual(string $searchKey, bool $tokenize = true): string
     {
         return self::GREATER_OR_EQUAL . trim($searchKey) . self::tokenize($tokenize);
     }
+
     /**
      * ...WHERE 1
      *    {{ AND searchKey < $value
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function lower(string $searchKey, bool $tokenize = true): string
     {
         return self::LOWER . trim($searchKey) . self::tokenize($tokenize);
     }
+
     /**
      * ...WHERE 1
      *    {{ AND searchKey <= $value }}
-     * 
-     * @param string $searchKey
+     *
      * @param bool $tokenize  if TRUE, a random hash is added to the returned string to ensure its uniqueness.
-     * @return string
      */
     public static function lowerOrEqual(string $searchKey, bool $tokenize = true): string
     {
         return self::LOWER_OR_EQUAL . trim($searchKey) . self::tokenize($tokenize);
     }
-    
+
+    /**
+     * @return iterable<string, string[]>
+     */
+    private static function getAlias(): iterable
+    {
+        yield self::EQUAL => ['=='];
+        yield self::NOT_EQUAL => ['!='];
+        yield self::OR => ['||'];
+    }
+
+    private static function getToken(): string
+    {
+        return bin2hex(random_bytes(15));
+    }
+
+    private static function tokenize(bool $truly): string
+    {
+        return $truly ? '~' . self::getToken() : '';
+    }
 }
